pyproject.toml
[tool.poetry]
name = "game-helper"
version = "0.1.0"
description = ""
authors = ["Алексей Дмитриев <xzenonoff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
isort = "^5.13.2"
ruff = "^0.3.5"
black = "^24.3.0"
djoser = "^2.2.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
django-filter = "^24.2"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.1"
nanoid = "^2.0.0"
dropbox = "^11.36.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
